backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# AI/ML Libraries
google-generativeai==0.3.2
groq==0.4.1
pinecone-client==3.0.2
tiktoken==0.5.2
numpy==1.26.3

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
requests==2.31.0

# Logging and monitoring
python-json-logger==2.0.7
